+++ pydatatask/repository/base.py
@@ -1,14 +1,7 @@
 """
-Repositories are arbitrary key-value stores. They are the data part of pydatatask.
-You can store your data in any way you desire and as long as you can write a Repository class to describe it, it can be
-used to drive a pipeline.
-
-The notion of the "value" part of the key-value store abstraction is defined very, very loosely. The repository base
-class doesn't have an interface to get or store values, only to query for and delete keys. Instead, you have to know
-which repository subclass you're working with, and use its interfaces. For example, `MetadataRepository` assumes that
-its values are structured objects and loads them fully into memory, and `BlobRepository` provides a streaming interface
-to a flat address space.
+This module contains repository base classes as well as repository combinators and the simple filesystem repositories.
 """
+
 from typing import (
     TYPE_CHECKING,
     Any,
@@ -18,7 +11,6 @@ from typing import (
     Callable,
     Coroutine,
     Dict,
-    List,
     Literal,
     Optional,
     overload,
@@ -33,16 +25,15 @@ import os
 import string
 
 import aiofiles.os
-import aiohttp.client_exceptions
 import aioshutil
 import yaml
 
+from ..utils import AReadStream, AReadText, AWriteStream, AWriteText, roundrobin
+
 if TYPE_CHECKING:
-    from ..task import ExecutorTask, KubeTask
+    from ..task import ExecutorTask
 
 l = logging.getLogger(__name__)
-
-from ..utils import AReadStream, AReadText, AWriteStream, AWriteText, roundrobin
 
 
 # Helper Functions
