+++ pydatatask/repository/mongodb.py
@@ -1,4 +1,7 @@
-from typing import TYPE_CHECKING, Any, Callable, Dict, Optional, overload
+"""
+This module contains repositories for interacting with MongoDB as a data store.
+"""
+from typing import Any, Callable, Dict, Optional
 
 import motor.core
 import motor.motor_asyncio
